pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-sql-sandbox"
version = "0.1.0"
description = "SQL/Data Cleaning Sandbox – An OpenEnv environment for agentic data engineering evaluation"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
inference = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "groq>=0.4.0",
]

[project.scripts]
server = "sql_sandbox.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["sql_sandbox", "sql_sandbox.server"]
package-dir = { "sql_sandbox" = ".", "sql_sandbox.server" = "server" }